simpleupload didn't fire *contentChange* event
simpleupload didn't trigger the *contentChange* event after the image upload success.
When I use ueditor with angular, I listen to the *contentChange* event to binding the data between editor and model.  When user upload an image, and then just save the model, the model only keep the placeholder image not the server response one.